Secondary change: extract map into SearchPageMap component
